fix: incorrect scope for haloCommentEnabled template variable
