+++ application/src/test/java/run/halo/app/theme/dialect/CommentElementTagProcessorTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.lenient;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import java.util.Map;
@@ -19,7 +18,6 @@ import org.thymeleaf.IEngineConfiguration;
 import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.context.Context;
 import org.thymeleaf.context.ITemplateContext;
-import org.thymeleaf.context.WebEngineContext;
 import org.thymeleaf.model.IProcessableElementTag;
 import org.thymeleaf.processor.element.IElementTagStructureHandler;
 import org.thymeleaf.spring6.dialect.SpringStandardDialect;
@@ -27,7 +25,6 @@ import org.thymeleaf.spring6.expression.ThymeleafEvaluationContext;
 import org.thymeleaf.templateresolver.StringTemplateResolver;
 import org.thymeleaf.templateresource.ITemplateResource;
 import org.thymeleaf.templateresource.StringTemplateResource;
-import org.thymeleaf.web.IWebExchange;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import run.halo.app.infra.SystemConfigurableEnvironmentFetcher;
@@ -103,74 +100,6 @@ class CommentElementTagProcessorTest {
               </body>
             </html>
             """);
-    }
-
-    @Test
-    void getCommentWidget() {
-        when(applicationContext.getBean(eq(SystemConfigurableEnvironmentFetcher.class)))
-            .thenReturn(environmentFetcher);
-        SystemSetting.Comment commentSetting = mock(SystemSetting.Comment.class);
-        when(environmentFetcher.fetchComment())
-            .thenReturn(Mono.just(commentSetting));
-
-        CommentWidget commentWidget = mock(CommentWidget.class);
-        when(extensionGetter.getEnabledExtensionByDefinition(CommentWidget.class))
-            .thenReturn(Flux.just(commentWidget));
-        WebEngineContext webContext = mock(WebEngineContext.class);
-        var evaluationContext = mock(ThymeleafEvaluationContext.class);
-        when(webContext.getVariable(
-            eq(ThymeleafEvaluationContext.THYMELEAF_EVALUATION_CONTEXT_CONTEXT_VARIABLE_NAME)))
-            .thenReturn(evaluationContext);
-        when(evaluationContext.getApplicationContext()).thenReturn(applicationContext);
-        IWebExchange webExchange = mock(IWebExchange.class);
-        when(webContext.getExchange()).thenReturn(webExchange);
-
-        // comment disabled
-        when(commentSetting.getEnable()).thenReturn(true);
-        assertThat(CommentElementTagProcessor.getCommentWidget(webContext).isPresent()).isTrue();
-
-        // comment enabled
-        when(commentSetting.getEnable()).thenReturn(false);
-        assertThat(CommentElementTagProcessor.getCommentWidget(webContext).isPresent()).isFalse();
-
-        // comment enabled and ENABLE_COMMENT_ATTRIBUTE is true
-        when(commentSetting.getEnable()).thenReturn(true);
-        when(webExchange.getAttributeValue(CommentWidget.ENABLE_COMMENT_ATTRIBUTE))
-            .thenReturn(true);
-        assertThat(CommentElementTagProcessor.getCommentWidget(webContext).isPresent()).isTrue();
-
-        // comment enabled and ENABLE_COMMENT_ATTRIBUTE is false
-        when(commentSetting.getEnable()).thenReturn(true);
-        when(webExchange.getAttributeValue(CommentWidget.ENABLE_COMMENT_ATTRIBUTE))
-            .thenReturn(false);
-        assertThat(CommentElementTagProcessor.getCommentWidget(webContext).isPresent()).isFalse();
-
-        // comment enabled and ENABLE_COMMENT_ATTRIBUTE is null
-        when(commentSetting.getEnable()).thenReturn(true);
-        when(webExchange.getAttributeValue(CommentWidget.ENABLE_COMMENT_ATTRIBUTE))
-            .thenReturn(null);
-        assertThat(CommentElementTagProcessor.getCommentWidget(webContext).isPresent()).isTrue();
-
-        // comment enabled and ENABLE_COMMENT_ATTRIBUTE is 'false'
-        when(commentSetting.getEnable()).thenReturn(true);
-        when(webExchange.getAttributeValue(CommentWidget.ENABLE_COMMENT_ATTRIBUTE))
-            .thenReturn("false");
-        assertThat(CommentElementTagProcessor.getCommentWidget(webContext).isPresent()).isFalse();
-    }
-
-    @Test
-    void populateAllowCommentAttribute() {
-        WebEngineContext webContext = mock(WebEngineContext.class);
-        IWebExchange webExchange = mock(IWebExchange.class);
-        when(webContext.getExchange()).thenReturn(webExchange);
-
-        CommentElementTagProcessor.populateAllowCommentAttribute(webContext, true);
-        verify(webExchange).setAttributeValue(
-            eq(CommentElementTagProcessor.COMMENT_ENABLED_MODEL_ATTRIBUTE), eq(true));
-
-        CommentElementTagProcessor.populateAllowCommentAttribute(webContext, false);
-        verify(webExchange).setAttributeValue(
-            eq(CommentElementTagProcessor.COMMENT_ENABLED_MODEL_ATTRIBUTE), eq(false));
     }
 
     static class DefaultCommentWidget implements CommentWidget {
